@@ -6,7 +6,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,27 +16,34 @@
 #
 
 source "https://rubygems.org"
+gem "rails", "4.0.2"
 
-gem "rails", "2.3.17"
-gem "rack", "1.1.6"
+gem "rack", "~> 1.5.2"
+gem "dotenv-rails", "~> 0.9.0"
+gem "haml-rails", "~> 0.5.3"
+gem "sass-rails", "~> 4.0.1"
+gem "jbuilder", "~> 2.0.3"
+gem "bcrypt", "~> 3.1.2"
+gem "kwalify", "~> 0.7.2"
+gem "simple-navigation", "~> 3.11.0"
+gem "sqlite3", "~> 1.3.8"
+gem "gli", "~> 2.9.0"
+gem "cocaine", "~> 0.5.3"
+gem "hashie", "~> 2.0.5"
+gem "activerecord-session_store", "~> 0.1.0"
+gem "chef", "~> 10.30.4"
 
-gem "app_config", "1.0.2"
-gem "chef", "10.24.4"
-gem "haml", "3.1.6"
-gem "hike", "1.2.1"
-gem "i18n", "0.4.2"
-gem "json", "1.6.1"
-gem "kwalify", "0.7.2"
-gem "multi_json", "1.0.3"
-gem "sass", "3.2.12"
-gem "simple-navigation", "3.7.0"
-gem "sprockets", "2.10.1"
-gem "sprockets-sass", "1.0.2"
-gem "sprockets-helpers", "1.1.0"
-gem "sqlite3", "1.3.6"
-gem "syslogger", "1.3.0"
-gem "tilt", "1.3.3"
-gem "mime-types", "1.18", :require => "mime/types"
+gem "uglifier", "~> 2.2.1"
+gem "therubyracer", "~> 0.12.1"
+
+# chef requires json <= 1.7.7 
+# rubocop requires json >= 1.7.7
+gem "json", "1.7.7"
+
+group :development, :test do
+  gem "debugger", "~> 1.6.5"
+  gem "rspec-rails", "~> 2.14.1"
+end
 
 # We are not adding any version here because the SLES and Ubuntu chef 
 # packaging differ for the version, but we need this to load the chef 
@@ -44,23 +51,25 @@ gem "mime-types", "1.18", :require => "mime/types"
 gem "chef"
 
 group :development do
-  gem "brakeman", "1.8.2"
-  gem "rspec-rails", "1.3.4"
+  gem "magic_encoding", "~> 0.0.2"
+  gem "guard", "~> 2.4.0"
+  gem "binding_of_caller", "~> 0.7.2"
+  gem "better_errors", "~> 1.1.0"
+  gem "meta_request", "~> 0.2.8"
+  gem "quiet_assets", "~> 1.0.2"
+  gem "pry-rails", "~> 0.3.2"
+  gem "brakeman", "~> 2.4.0", require: false
+  gem "rubocop", "~> 0.18.1", require: false
 end
 
 group :test do
-  gem "rspec-rails", "1.3.4"
-  gem "mocha", "0.10.5"
-  gem "sinatra", "1.2.9"
-  gem "webmock", "1.16.1"
-  gem "rcov", "1.0.0"
+  gem "mocha", "~> 1.0.0"
+  gem "webmock", "~> 1.17.3"
+  gem "sinatra", "~> 1.4.4"
+  gem "cucumber-rails", "~> 1.4.0", require: false
+  gem "simplecov", "~> 0.8.2", require: false
 end
 
-# Install gems from each barclamp. The crowbar_path variable is
-# needed by Gemfile.d/*.gemfile so that they can determine the correct
-# path at run-time rather than having to hard-code a path at
-# install-time which may be different and therefore incorrect.
-crowbar_path = File.expand_path("..", __FILE__)
-Dir.glob(File.join(crowbar_path, "Gemfile.d", "*.gemfile")) do |gemfile|
-  eval(IO.read(gemfile), binding)
+group :doc do
+  gem "sdoc", "~> 0.3.20", require: false
 end
