@@ -1,11 +1,12 @@
+#
 # Copyright 2011-2013, Dell
-# Copyright 2013, SUSE LINUX Products GmbH
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#   http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,105 +14,121 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Rob Hirschfeld
-# Author: SUSE LINUX Products GmbH
-#
 
-ActionController::Routing::Routes.draw do |map|
-  # Install route from each barclamp
-  Dir.glob(File.join(File.dirname(__FILE__), 'routes.d', '*.routes')) do |routes_file|
-      eval(IO.read(routes_file), binding)
+Crowbar::Application.routes.draw do
+  root :to => "dashboard#index"
+
+  Dir.glob(File.join(File.dirname(__FILE__), "routes.d", "*.routes")) do |routes_file|
+    eval(IO.read(routes_file), binding)
   end
 
-  map.root :controller => "nodes", :action=>'index'
+  constraints(id: /.*/ ) do
+    get "dashboard" => "dashboard#index", as: :dashboard
+    get "dashboard/:id" => "dashboard#show", as: :dashboard_detail
 
-  # documentation / help
-  map.docs 'docs', :controller=>'docs', :action=>'index', :conditions => { :method => :get }
-  map.docs_topic 'docs/topic/:id', :controller=>'docs', :action=>'topic', :conditions => { :method => :get }
-  map.docs_barclamp 'docs/:controller/:id', :action=>'docs', :conditions => { :method => :get }
+    resources :nodes, only: [:edit, :update, :show] do
+      collection do
+        get :status
+        get :unallocated
+        get :list
+        get :families
+        post :bulk
+      end
 
-  # nodes
-  map.resources :nodes, :only => [:index]
-  map.connect 'nodes/:name/attribute/*path', :controller => 'nodes', :action => 'attribute',
-              :requirements => { :name => /.*/, :path => /.*/ }, :conditions => { :method => :get }
-  map.nodes_status 'nodes/status.:format', :controller => 'nodes', :action => 'status', :conditions => { :method => :get }
-  map.nodes_list 'nodes/list.:format', :controller => 'nodes', :action => 'list', :conditions => { :method => :get }
-  map.unallocated_list 'nodes/unallocated.:format', :controller => 'nodes', :action => 'unallocated', :conditions => { :method => :get }
-  map.bulk_nodes 'nodes/bulk.:format', :controller => 'nodes', :action => 'bulk', :conditions => { :method => :post }
-  map.nodes_families 'nodes/families', :controller=>'nodes', :action=>'families'
-  map.hit_node 'nodes/:id/hit/:req', :controller=>'nodes', :action=>'hit', :requirements => { :id => /.*/ }
-  map.edit_node 'nodes/:name/edit', :controller=>'nodes', :action =>'edit', :requirements => { :name => /.*/ }
-  map.dashboard 'dashboard', :controller => 'nodes', :action => 'index'
-  map.dashboard_detail 'dashboard/:name', :controller => 'nodes', :action => 'index', :requirements => { :name => /.*/ }
-  map.group_change 'nodes/groups/1.0/:id/:group', :controller => 'nodes', :action=>'group_change', :conditions => { :method => :post }, :requirements => { :id => /.*/ }
-  # this route allows any barclamp to extend the nodes view
-  map.nodes_barclamp 'nodes/:controller/1.0', :action => 'nodes'  
-  map.update_node 'nodes/:name/update', :controller => 'nodes', :action=>'update', :requirements => { :name => /.*/ }
-  map.node 'nodes/:name', :controller => 'nodes', :action => 'show', :requirements => { :name => /.*/ }
-  
-  # this route allows any barclamp to extend the network view
-  map.network_barclamp 'network/:controller/1.0', :action=>'network'
-  # these paths require the network barclamp 
-  map.network 'network', :controller => 'network', :action=>'switch'
-  map.switch 'network/switch/:id', :controller => 'network', :action=>'switch', :requirements => { :id => /.*/ }
-  map.vlan 'network/vlan/:id', :controller => 'network', :action=>'vlan', :requirements => { :id => /.*/ }
-  
-  #support paths
+      member do
+        get :status
+        post :group
 
+        get "hit/:req" => "nodes#hit", as: :hit
+        get "attribute/*path" => "nodes#attribute", as: :attribute, constraints: { path: /.*/ }
+      end
+    end
 
-  map.utils 'utils.:format', :controller=>'support', :action=>'index'
-  map.utils_files 'utils/files/:id', :controller=>'support', :action=>'destroy', :requirements => { :id => /.*/ }
-  map.export_chef 'utils/chef', :controller=>'support', :action=>'export_chef'
-  map.export_supportconfig 'utils/supportconfig', :controller=>'support', :action=>'export_supportconfig'
-  map.utils_export 'utils/:controller/1.0/export', :action=>'export'
-  map.utils_barclamp 'utils/:controller/1.0', :action=>'utils'
-  map.utils_import 'utils/import/:id', :controller=>'support', :action=>'import', :requirements => { :id => /.*/ }
-  map.utils_upload 'utils/upload/:id', :controller=>'support', :action=>'upload', :requirements => { :id => /.*/ }
-  map.restart 'utils/restart/:id.:format', :controller=>'support', :action=>'restart'
-  
-  # barclamps
-  map.help_barclamp             'crowbar/:controller/1.0/help', :action => 'help', :conditions => { :method => :get }
-  map.barclamp_nodes            'crowbar/:controller/1.0/proposals/nodes', :action=>'nodes'
-  map.create_proposal_barclamp  'crowbar/:controller/1.0/proposals', :action => 'proposal_create', :conditions => { :method => :put }
-  map.proposals_barclamp        'crowbar/:controller/1.0/proposals', :action => 'proposals', :conditions => { :method => :get }
-  map.commit_proposal_barclamp  'crowbar/:controller/1.0/proposals/commit/:id', :action => 'proposal_commit', :conditions => { :method => :post }
-  map.status_proposals_barclamp 'crowbar/:controller/1.0/proposals/status/:id/.:format', :action => 'proposal_status', :conditions => { :method => :get }
-  map.delete_proposal_barclamp  'crowbar/:controller/1.0/proposals/:id', :action => 'proposal_delete', :conditions => { :method => :delete }
-  map.dequeue_barclamp  'crowbar/:controller/1.0/proposals/dequeue/:id', :action => 'proposal_dequeue', :conditions => { :method => :delete }
-  map.update_proposal_barclamp  'crowbar/:controller/1.0/proposals/:id', :action => 'proposal_update', :conditions => { :method => :post }
-  map.proposal_barclamp         'crowbar/:controller/1.0/proposals/:id', :action => 'proposal_show', :conditions => { :method => :get }
-  map.connect 'crowbar/:controller/1.0/elements', :action => 'elements', :conditions => { :method => :get }
-  map.connect 'crowbar/:controller/1.0/elements/:id', :action => 'element_info', :conditions => { :method => :get }
-  map.connect 'crowbar/:controller/1.0/transition/:id', :action => 'transition', :conditions => { :method => :post }
-  map.connect 'crowbar/:controller/1.0/transition/:id', :action => 'transition', :conditions => { :method => :get }
-  map.index_barclamp            'crowbar/:controller/1.0', :action => 'index', :conditions => { :method => :get }
-  map.delete_barclamp           'crowbar/:controller/1.0/:id', :action => 'delete', :conditions => { :method => :delete }, :requirements => { :id => /.*/ }
-  map.show_barclamp             'crowbar/:controller/1.0/:id', :action => 'show', :conditions => { :method => :get }, :requirements => { :id => /.*/ }
-  map.versions_barclamp         'crowbar/:controller', :action => 'versions', :conditions => { :method => :get }
-  map.action_barclamp           'crowbar/:controller/1.0/:action/:id', :conditions => { :method => :post }, :requirements => { :id => /.*/ }
-  map.barclamp_index_barclamp   'crowbar', :controller => 'barclamp', :action => 'barclamp_index', :conditions => { :method => :get }
-  map.barclamp_modules 'crowbar/modules/1.0', :controller => 'barclamp', :action => 'modules', :conditions => { :method => :get }
+    resources :utils, only: [:index, :destroy], controller: :support do
+      collection do
+        get :chef # maybe move into custom export controller? export_barclamp_path(controller: "chef")
+        get :supportconfig # maybe move into custom export controller? export_barclamp_path(controller: "supportconfig")
+      end
 
-  map.connect 'crowbar/:barclamp/1.0/help', :action => 'help', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals/nodes', :controller => "barclamp", :action=>'nodes'
-  map.connect 'crowbar/:barclamp/1.0/proposals', :action => 'proposal_create', :conditions => { :method => :put }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals', :action => 'proposals', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals/commit/:id', :action => 'proposal_commit', :conditions => { :method => :post }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals/status.:format', :action => 'proposal_status', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals/:id', :action => 'proposal_delete', :conditions => { :method => :delete }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals/:id', :action => 'proposal_update', :conditions => { :method => :post }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/proposals/:id', :action => 'proposal_show', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/elements', :action => 'elements', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/elements/:id', :action => 'element_info', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/transition/:id', :action => 'transition', :conditions => { :method => :post }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/transition/:id', :action => 'transition', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0', :action => 'index', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/status', :action => 'status', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/:id', :action => 'delete', :conditions => { :method => :delete }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/:id', :action => 'show', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp', :action => 'versions', :conditions => { :method => :get }, :controller => 'barclamp'
-  map.connect 'crowbar/:barclamp/1.0/:action/:id', :conditions => { :method => :post }, :controller => 'barclamp'
+      member do
+        get :restart
+        post :import
+        post :upload
+      end
+    end
 
-  map.connect ':controller/:action/:id'
-  map.connect ':controller/:action/:id.:format'
+    resources :networks, only: [] do
+      member do 
+        get :switch
+        get :vlan
+      end
+    end
+
+    resources :docs, only: [:index] do
+      collection do
+        get "*id" => "docs#show"
+      end
+    end
+  end
+
+  match "crowbar/modules/1.0", controller: "barclamp", action: "modules", as: :crowbar_barclamps, via: [:get, :post]
+
+  match "nodes/:controller/1.0", action: "nodes", as: :nodes_barclamp, via: [:get, :post]
+  match "utils/:controller/1.0", action: "utils", as: :utils_barclamp, via: [:get, :post]
+  match "export/:controller/1.0", action: "export", as: :export_barclamp, via: [:get, :post]
+  match "network/:controller/1.0", action: "network", as: :network_barclamp, via: [:get, :post]
+  match "docs/:controller/1.0", action: "docs", as: :docs_barclamp, via: [:get, :post]
+
+  constraints(id: /.*/ ) do
+    get "crowbar/:controller/1.0", action: :index, as: :barclamps
+    get "crowbar/:controller/1.0/:id", action: :show, as: :barclamp_show
+    delete "crowbar/:controller/1.0/:id", action: :delete, as: :barclamp_delete
+    post "crowbar/:controller/1.0/:action/:id", as: :barclamp_action
+  end
+
+  scope "crowbar/:controller/1.0" do
+    match "help", action: "help", as: :help, via: [:get]
+
+    match "proposals", action: "proposals", as: :proposals, via: [:get]
+    match "proposals", action: "proposal_create", as: :proposal_create, via: [:put]
+    match "proposals/:id", action: "proposal_update", as: :proposal_update, via: [:post]
+    match "proposals/:id", action: "proposal_show", as: :proposal_show, via: [:get]
+    match "proposals/delete/:id", action: "proposal_delete", as: :proposal_delete, via: [:get]
+    match "proposals/dequeue/:id", action: "proposal_dequeue", as: :proposal_dequeue, via: [:get]
+    match "proposals/status/:id", action: "proposal_status", as: :proposal_status, via: [:get]
+    match "proposals/commit/:id", action: "proposal_commit", as: :proposal_commit, via: [:post]
+
+    match "elements", action: "elements", as: :elements, via: [:get]
+    match "elements/:id", action: "element_info", as: :element, via: [:get]
+
+    match "transition/:id", action: "transition", as: :transition, via: [:get, :post]
+  end
+  get "crowbar/:controller", action: "versions"
+
+  constraints(id: /.*/ ) do
+    get "crowbar/:barclamp/1.0", controller: "barclamp", action: :index
+    get "crowbar/:barclamp/1.0/:id", controller: "barclamp", action: :show
+    delete "crowbar/:barclamp/1.0/:id", controller: "barclamp", action: :delete
+    post "crowbar/:barclamp/1.0/:action/:id", controller: "barclamp"
+  end
+
+  scope "crowbar/:barclamp/1.0", controller: "barclamp" do
+    match "help", action: "help", via: [:get]
+
+    match "proposals", action: "proposals", via: [:get]
+    match "proposals", action: "proposal_create", via: [:put]
+    match "proposals/:id", action: "proposal_update", via: [:post]
+    match "proposals/:id", action: "proposal_show", via: [:get]
+    match "proposals/delete/:id", action: "proposal_delete", via: [:get]
+    match "proposals/dequeue/:id", action: "proposal_dequeue", via: [:get]
+    match "proposals/status/:id", action: "proposal_status", via: [:get]
+    match "proposals/commit/:id", action: "proposal_commit", via: [:post]
+
+    match "elements", action: "elements", via: [:get]
+    match "elements/:id", action: "element_info", via: [:get]
+
+    match "transition/:id", action: "transition", via: [:get, :post]
+  end
+  get "crowbar/:barclamp", controller: "barclamp", action: "versions" 
 end
+
