@@ -1,68 +1,55 @@
-# Copyright 2011, Dell 
-# 
-# Licensed under the Apache License, Version 2.0 (the "License"); 
-# you may not use this file except in compliance with the License. 
-# You may obtain a copy of the License at 
-# 
-#  http://www.apache.org/licenses/LICENSE-2.0 
-# 
-# Unless required by applicable law or agreed to in writing, software 
-# distributed under the License is distributed on an "AS IS" BASIS, 
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
-# See the License for the specific language governing permissions and 
-# limitations under the License. 
-# 
-# Author: RobHirschfeld 
-# 
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
 
 require 'uri'
 
-# Filters added to this controller apply to all controllers in the application.
-# Likewise, all the methods added will be available for all controllers.
+
+
 class ApplicationController < ActionController::Base
+  protect_from_forgery with: :exception
+  helper :all
+
+
 
   @@users = nil
   
   before_filter :digest_authenticate, :if => :need_to_auth?
-  
-  
-  # Basis for the reflection/help system.
-  
-  # First, a place to stash the help contents.  
-  # Using a class_inheritable_accessor ensures that 
-  # these contents are inherited by children, but can be 
-  # overridden or appended to by child classes without messing up 
-  # the contents we are building here.
-  class_inheritable_accessor :help_contents
-  self.help_contents = []
-  
-  # Class method for adding method-specific help/API information 
-  # for each method we are going to expose to the CLI.
-  # Since it is a class method, it will not be bothered by the Rails
-  # trying to expose it to everything else, and we can call it to build
-  # up our help contents at class creation time instead of instance creation
-  # time, so there is minimal overhead.
-  # Since we are just storing an arrray of singleton hashes, adding more
-  # user-oriented stuff (descriptions, exmaples, etc.) should not be a problem.
-  def self.add_help(method,args=[],http_method=[:get])
-    # if we were passed multiple http_methods, build an entry for each.
-    # This assumes that they all take the same parameters, if they do not
-    # you should call add_help for each different set of parameters that the
-    # method/http_method combo can take.
-    http_method.each { |m|
-      self.help_contents = self.help_contents.push({
-        method => {
-                                             "args" => args,
-                                             "http_method" => m
-        }
-      })
-    }
+
+  class << self
+    def help_contents
+      @help_contents ||= []
+    end
+
+    def add_help(method, args=[], http_method = [:get])
+      http_method.each { |m|
+        help_contents.push({
+          method => {
+            "args" => args,
+            "http_method" => m
+          }
+        })
+      }
+    end
   end
+
   
-  helper :all
-  
-  protect_from_forgery # See ActionController::RequestForgeryProtection for details
-  
+
   def self.set_layout(template = "application")
     layout proc { |controller| 
       if controller.is_ajax? 
@@ -184,3 +171,4 @@ class ApplicationController < ActionController::Base
     end
   end
 end
+
