@@ -31,21 +31,23 @@ module ApplicationHelper
     @@suse_system ||= File.exist?("/etc/SuSE-release")
   end
 
-  # Added this helper method to access app config, maybe we need some wrapping
-  # and it feels better to call a method instead of a class
-  def app_config
-    AppConfig
-  end
-
   # A simple wrapper to access the branding configuration directly, looks much
   # cleaner within the views
   def branding_config
-    app_config[:branding]
+    @branding_config ||= begin
+      config = YAML::load_file(
+        Rails.root.join("config", "branding.yml")
+      )
+
+      Hashie::Mash.new config
+    rescue
+      Hashie::Mash.new
+    end
   end
 
   # Generate the meta title that gets displayed on the page meta information
   def meta_title
-    title = [branding_config[:page_title], branding_config[:page_slogan]].compact.join(" ")
+    title = [branding_config.page_title, branding_config.page_slogan].compact.join(" ")
     "#{title}: #{controller.action_name.titleize}"
   end
 
@@ -86,18 +88,18 @@ module ApplicationHelper
     [].tap do |output|
       output.push content_tag(
         :span,
-        branding_config[:page_title],
+        branding_config.page_title,
         :class => "title"
       )
 
-      unless branding_config[:page_slogan].empty?
+      unless branding_config.page_slogan.empty?
         output.push content_tag(
           :span,
-          branding_config[:page_slogan],
+          branding_config.page_slogan,
           :class => "slogan"
         )
       end
-    end.join("\n")
+    end.join("\n").html_safe
   end
 
   # Include required meta tags like csrf token, viewport and such stuff
@@ -119,21 +121,7 @@ module ApplicationHelper
         :name => "viewport",
         :content => "width=device-width, initial-scale=1.0"
       )
-
-      if protect_against_forgery?
-        output.push tag(
-          :meta,
-          :name => "csrf-param",
-          :content => Rack::Utils.escape_html(request_forgery_protection_token)
-        )
-
-        output.push tag(
-          :meta,
-          :name => "csrf-token",
-          :content => Rack::Utils.escape_html(form_authenticity_token)
-        )
-      end
-    end.join("\n")
+    end.join("\n").html_safe
   end
 
   def have_openstack
