@@ -1,30 +1,26 @@
 #
-# Copyright 2011, Dell
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
-#     http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: andi abes
-#
-
-####
-# if monitored by nagios, install the nrpe commands
 
 # Node addresses are dynamic and can't be set from attributes only.
 node[:==BC-MODEL==][:monitor][:ports]["==BC-MODEL==-api"] = [node[:==BC-MODEL==][:api_bind_host], node[:==BC-MODEL==][:api_bind_port]]
 
 svcs = node[:==BC-MODEL==][:monitor][:svcs]
 ports = node[:==BC-MODEL==][:monitor][:ports]
-log ("will monitor ==BC-MODEL== svcs: #{svcs.join(',')} and ports #{ports.values.join(',')}")
+log("will monitor ==BC-MODEL== svcs: #{svcs.join(',')} and ports #{ports.values.join(',')}")
 
 include_recipe "nagios::common" if node["roles"].include?("nagios-client")
 
@@ -36,7 +32,11 @@ template "/etc/nagios/nrpe.d/==BC-MODEL==_nrpe.cfg" do
   variables( {
     :svcs => svcs ,
     :ports => ports
-  })    
-   notifies :restart, "service[nagios-nrpe-server]"
-end if node["roles"].include?("nagios-client")    
+  })
+  notifies :restart, "service[nagios-nrpe-server]"
+
+  only_if do
+    node["roles"].include?("nagios-client")
+  end
+end
 
