@@ -1,6 +1,18 @@
 #
-# Cookbook Name:: ==*BC-MODEL==
-# Recipe:: setup
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 
 include_recipe "#{@cookbook_name}::common"
@@ -13,5 +25,9 @@ bash "tty linux setup" do
 	curl #{node[:==BC-MODEL==][:tty_linux_image]} | tar xvz -C /tmp/
 	touch /var/lib/==BC-MODEL==/tty_setup
   EOH
-  not_if do File.exists?("/var/lib/==BC-MODEL==/tty_setup") end
+
+  not_if do 
+    File.exists?("/var/lib/==BC-MODEL==/tty_setup")
+  end
 end
+
