@@ -1,9 +1,28 @@
+#
+# Copyright 2011-2013, Dell
+# Copyright 2013-2014, SUSE LINUX Products GmbH
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 name "==BC-MODEL==-server"
-description "==*BC-MODEL== Server Role"
+description "==*BC-MODEL== server role"
+
 run_list(
-         "recipe[==BC-MODEL==::api]",
-         "recipe[==BC-MODEL==::monitor]"
+  "recipe[==BC-MODEL==::api]",
+  "recipe[==BC-MODEL==::monitor]"
 )
+
 default_attributes()
 override_attributes()
 
